Gestion d'erreurs v2

+++ emapp/src/main/java/se/m1/emapp/model/core/JPAManager.java
@@ -9,13 +9,11 @@ import java.util.ArrayList;
 import java.util.List;
 import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
-import javax.persistence.EntityNotFoundException;
-import javax.persistence.NoResultException;
 import javax.persistence.PersistenceContext;
 import javax.persistence.TypedQuery;
 import se.m1.emapp.model.business.Credential;
 import se.m1.emapp.model.business.Employee;
-
+import javax.ejb.EJBException;
 @Stateless
 public class JPAManager {
 
@@ -31,36 +29,53 @@ public class JPAManager {
     /**
      * allows the creation of an employee
      * @param e the employee to create
+     * @throws EJBException
      */
-    public void createEmployee(Employee e)throws RuntimeException{ 
+    public void createEmployee(Employee e)  throws EJBException{ 
         em.persist(e);
     }
     
     /**
      * allows employee deletion
      * @param e the employee to delete
+     * @return a boolean that confirms if the employee is deleted
+     * @throws EJBException
      */
-    public void removeEmployee(Employee e)throws RuntimeException{
-        if (!em.contains(e)){
-            e = em.merge(e);
+    public boolean removeEmployee(Employee e)  throws EJBException{
+        if(read(e.getId())==null){
+            return false;
+        }else{
+            if (!em.contains(e)){
+                e = em.merge(e);
+            }
+            em.remove(e);
+            return true;
         }
-        em.remove(e);
+
     }
     
     /**
      * allows the modification of an employee
      * @param e the employee to modify
+     * @return a boolean that confirms if the employee is modified
+     * @throws EJBException
      */
-    public void modifyEmployee(Employee e) throws RuntimeException{ 
+    public boolean modifyEmployee(Employee e)  throws EJBException{ 
         em.getEntityManagerFactory().getCache().evictAll();
-        em.merge(e);
+        if(read(e.getId())==null){
+            return false;
+        }else{
+            em.merge(e);
+            return true;
+        }
     }
     
     /**
      * Allows to get all the employee
      * @return an ArrayList with all the employees
+     * @throws EJBException
      */
-    public List<Employee> getAll() throws RuntimeException{ 
+    public List<Employee> getAll()  throws EJBException{ 
         queryEmployee = em.createNamedQuery("Employee.findAll", Employee.class);
         all = new ArrayList<>(queryEmployee.getResultList());
         return all;      
@@ -70,24 +85,28 @@ public class JPAManager {
      * get the details of an employee
      * @param id of the employee to read
      * @return The employee corresponding to the id
+     * @throws EJBException
      */
-    public Employee read(int id){
-            em.getEntityManagerFactory().getCache().evictAll();
-            queryEmployee = em.createNamedQuery("Employee.findById", Employee.class);
-            queryEmployee.setParameter("id",id);         
-            if (queryEmployee.getResultList().isEmpty()){
-                return null;
-            }
-            Employee x = queryEmployee.getSingleResult();
-            return x;
+       
+    public Employee read(int id) throws EJBException{
+        em.getEntityManagerFactory().getCache().evictAll();
+        queryEmployee = em.createNamedQuery("Employee.findById", Employee.class);
+        queryEmployee.setParameter("id",id);
+        //we verify if the employee exists :
+        if (queryEmployee.getResultList().isEmpty()){
+            return null;
+        }
+        Employee x = queryEmployee.getSingleResult();
+        return x;
     }
     
     /**
      * verify the credentials entered
      * @param c the credentials to be verified
      * @return a boolean : true if the credentials entered are valid, false if not
+     * @throws EJBException
      */
-    public boolean checkCredentials(Credential c) throws RuntimeException{
+    public boolean checkCredentials(Credential c)  throws EJBException{
         queryCredential = em.createNamedQuery("Credential.checkcred", Credential.class);
         queryCredential.setParameter("login",c.getLogin());
         queryCredential.setParameter("pwd",c.getPwd());
